feat(llm): generalize the chat client into rag-rat-llm with guided decoding
Move the single-turn chat client out of rag-rat-dream/src/model.rs into a new
rag-rat-llm `chat` module, generalizing the names: `ChatModel` trait,
`HttpChatModel`, `provision_chat_model`, and the request/response types. Nothing
in the client was dream-specific — the verdict-only logic (prompt render,
citation guard, parse) stays in dream's `verdict`/`compact`.

Add backend-aware guided decoding: the trait gains `complete_guided(prompt,
Option<GuidedJson>)` (required), with `complete(prompt)` defaulting to the
unguided path so existing callers are unchanged. A vLLM client emits
`response_format: {type: json_schema}`, an Ollama client emits `format:
<schema>`, and any other backend degrades to unguided.

Migrate every call site to the new location with no re-export shim: dream's
verdict/compact passes take `&dyn ChatModel`, the `dream` CLI command builds via
`rag_rat_llm::chat::{HttpChatModel, provision_chat_model}`, and the test mock
relocates to a dream test-support module (`MockChatModel`). Drop the now-unused
`ureq` dependency from rag-rat-dream. Share the loopback-classification helper
with the embedder client rather than duplicating it.

Prerequisite for the distill LLM pass (#704). Part of #113. Closes #769.
