+++ crates/rag-rat-dream/Cargo.toml
@@ -24,7 +24,6 @@ serde_json.workspace = true
 sha2.workspace = true
 strum.workspace = true
 tracing.workspace = true
-ureq.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
